Tarefa1_2022li1g020

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

mapaValido :: Mapa -> Bool

Função Principal

Expand

A função mapaValido verifica se um certo mapa não viola certas restrições, que são invocadas através do uso de funções auxiliares.

Exemplos de utilização:

>>> mapaValido (Mapa 2 ((Rio,(Nenhum,Tronco)),(Relva,(Nenhum,Nenhum))))

True

Propriedades:

mapaValido (Mapa larg []) = []

mapaValido1 :: Mapa -> Bool

Funções Auxiliares

Expand

Função Auxiliar 1

A função auxiliar mapaValido1 verifica se não existem obstáculos em terrenos impróprios, como por exemplo, troncos em estradas ou relvas, ́arvores em rios ou estradas e carros em relvas e rios.

Exemplos de utilização:

>>> mapaValido1 (Mapa 2 (Relva,(Nenhum,Carro)))
False   

>>> mapaValido1 (Mapa 2 (Rio,(Tronco,Nenhum)))
True

Propriedades:

mapaValido1 (Mapa 3 []) = True

mapaValido2 :: Mapa -> Bool

Função Auxiliar 2

A função auxiliar mapaValido2 verifica se Rios contíguos têm direções opostas, ou seja, rios que apareçam seguidos não podem apresentar a mesma direção. Neste caso utilizamos outra função (mapaAuxRios) para quando dois terrenos seguidos sejam iguais, esta descarte se não forem rios e se forem execute.se de forma a que a multiplicação das suas velocidades seja menor que zero. Ao dizer isto, consideramos uma velocidade negativa e outra positiva, o que faz com que os rios apresentem direções contrárias.

Exemplos de utilização:

>>> mapaValido2 (Mapa 2 (Relva,(Nenhum,Carro),(Rio 3,(Nenhum,Tronco))))
True 

>>> mapaValido2 (Mapa 2 (Rio 2,(Tronco,Nenhum)),(Rio 2,(Tronco,Nenhum)))
False

Propriedades:

mapaValido2 (Mapa 3 []) = True

mapaValido3 :: Mapa -> Bool

Função Auxiliar 3

A função auxiliar mapaValido3 verifica se quando estamos num rio, troncos têm, no máximo, 5 unidades de comprimento. Para isto acontecer usamos uma função auxiliar (/mapaAuxTronco) que verifica, quando o rio tem mais do que 5 unidades de comprimento, se existe um tronco maior do que 5, ou seja, que mais de 5 obstáculos seguidos sejam Tronco.

Exemplos de utilização:

>>> mapaValido3 (Mapa 3 (Rio 3,(Tronco,Tronco,Nenhum)))
True

>>> mapaValido3 (Mapa 7 (Rio 5,(Tronco,Tronco,Tronco,Tronco,Tronco,Tronco,Nenhum)))
False

Propriedades:

mapaValido3 (Mapa largura (Rio 5,[]) = True

auxiliar3aux :: [Obstaculo] -> Int -> Bool

mapaValido4 :: Mapa -> Bool

Função Auxiliar 4

A função auxiliar mapaValido4 funciona quase da mesma forma que a função anterior, só que neste caso serve para verificar se os carros não tem mais de 3 unidades de comprimento. Para isso usamos uma função auxiliar (mapaAuxCarros) que, quando a estrada tiver mais de 3 unidades de comprimento, vê se existem mais de 3 obstáculos seguido com o nome Carro.

Exemplos de utilização:

>>> mapaValido4 (Mapa 3 (Estrada 3,(Carro,Carro,Nenhum)))
True

>>> mapaValido4 (Mapa 5 (Rio 6,(Carro,Carro,Carro,Carro,Nenhum,Nenhum)))
False

Propriedades:

mapaValido4 (Mapa largura (Estrada 3,[]) = True

mapaValido4aux :: [Obstaculo] -> Int -> Bool

mapaValido5 :: Mapa -> Bool

Função Auxiliar 5

A função auxiliar mapaValido5 observamos se existe pelo menos em cada linha, um obstáculo Nenhum. Isto serve para o jogador poder atravessar essa mesma linha enquanto decorre o jogo.

Exemplos de utilização:

>>> mapaValido5 (Mapa 3 (Estrada -1,(Carro,Carro,Nenhum)))
True

>>> mapaValido5 (Mapa 4 (Estrada 3,(Carro,Carro,Carro,Carro,)))
False

Propriedades:

mapaValido5 (Mapa largura []) = True

mapaValido6 :: Mapa -> Bool

Função Auxiliar 6

A função auxiliar mapaValido6 faz com que o comprimento da lista de obstáculos corresponda exatamente à largura do mapa. === Exemplos de utilização:

>>> mapaValido6 (Mapa 3 (Estrada 3,(Carro,Carro,Nenhum)))
True

>>> mapaValido6 (Mapa 5 (Rio -2,(Tronco,Nenhum,Nenhum)))
False

Propriedades:

mapaValido6 (Mapa largura []) = True

mapaValido7 :: Mapa -> Bool

Função Auxiliar 7

A função auxiliar mapaValido7 existe para verificar se não existem de forma seguida no mapa, mais do que 4 rios, nem 5 estradas ou relvas.

Exemplos de utilização:

>>> mapaValido7 (Mapa 3 (Estrada 3,(Carro,Carro,Nenhum),(Estrada 3,(Carro,Carro,Nenhum),(Rio 6,(Nenhum,Nenhum,Tronco)))
True

>>> mapaValido7 (Mapa 2 (Rio 6,(´Tronco,Nenhum),(Rio -3,(Tronco,Nenhum),(Rio 4,(Tronco,Nenhum),(Rio -2,(Tronco,Nenhum),(Rio 3,(Tronco,Nenhum)))
False

Propriedades:

mapaValido7 (Mapa largura []) = True

mapaValido8 :: Largura -> [(Terreno, [Obstaculo])] -> Bool