Tarefa2_2022li1g020

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/23.

Documentation

estendeMapa :: Mapa -> Int -> Mapa

Função principal

A função estendeMapa deve gerar e adicionar uma nova linha válida ao topo de um dado mapa. Para isso é necessário utilizar funções auxiliares.

Propriedades:

estendeMapa m 0 = m

estendeMapa1 :: Mapa -> Int -> Mapa

criaobs :: Mapa -> Int -> [Obstaculo]

numeroRandom :: Int -> Int -> [Int]

A função numeroRandom a partir de deois números inteiros vai gerar números grandes aleatórios que vão contribuir na expansão do mapa que será dado e que constitui o jogo.

proximosTerrenosValidos :: Mapa -> [Terreno]

Funções Auxiliares

Expand

Função Auxiliar 1

A função proximosTerrenosValidos deve gerar a lista de terrenos passíveis de serem usados numa nova linha no topo do mapa dado. Assumimos para esta funçao que a velocidade é 0. Utilizamos uma função auxiliar por acumuladores, pois temos de ter em conta o aspeto de mais do que 4 rios, nem 5 estradas ou relvas.

Exemplos de utilização:

>>> proximosTerrenosValidos (Mapa 2 [(Rio 0,[Nenhum,Arvore,Carro])])
[Rio 0,Estrada 0,Relva]

>>> proximosTerrenosValidos (Mapa _ [(Rio _, _),(Rio _, _),(Rio _, _),(Rio _, _),_])
[Estrada 0, Relva]

proximosObstaculosValidos :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

Função Auxiliar 2

Expand

A função proximosObstaculosValidos deve gerar a lista de obstaculos passiveis de serem usados para continuar uma dada linha do mapa.

Nesta função, deve ter em atenção não só os obstáculos permitidos no tipo de terreno indicado como as regras a respeito do comprimento dos obstaculos.

Exemplos de utilização:

>>> proximosObstaculosValidos (Mapa 2 [(Estrada 3,[Carro,Nenhum,Arvore,Nenhum])])
False

>>> proximosObstaculosValidos (Mapa 3 [(Rio 2,[Carro,Nenhum,Arvore,Nenhum])])
False

Propriedades:

proximosObstaculosValidos a [] = []

observaTerreno :: Terreno -> Int

A função "observaTerreno" recebe um terreno e devolve um inteiro, que é utilizado para substituir o nome do terreno presente.

contaTroncos :: [Obstaculo] -> Int -> [Obstaculo]

As funções contaTroncos e contaCarros recebem uma lista de obstaculos e um número inteiro até atingit o limite máximo possível (5 e 3 respetivmente). O número dado vai aumentando se encontrar esse obstaculo.

contaCarros :: [Obstaculo] -> Int -> [Obstaculo]