Tarefa3_2022li1g020

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2022/23.

Documentation

animaJogo :: Jogo -> Jogada -> Jogo

A função animaJogo movimenta os obstáculos (em função da velocidade) do terreno em que se encontram e o personagem de acordo com a jogada dada.

moveObstaculos :: [(Terreno, [Obstaculo])] -> [Int]

Funções 1

A função moveObstaculos tem em conta de que numa estrada ou rio com velocidade v, os obstáculos devem mover-se |v| unidades na direcção determinada. No rio, a velocidade do mesmo tem de ser a mesma que a do tronco (obstaculo). No outro caso, a velocidade da estrada terá que ser a mesma que a do carro.

Propriedades:

moveObstaculos [] = []

moveJogador :: Largura -> Coordenadas -> Jogada -> Coordenadas

Função 2

A função moveJogador dita as coordenadas do jogador quando é feita uma jogada pelo jogador. Tendo em conta uma função definida seguidamente.

Exemplos de utilização:

>>> moveJogador (3,-2) Move Cima
(3,-3)

movimentoNoTronco :: Jogador -> Mapa -> Jogador

Função 3

A função movimentoNoTronco terá que ter a mesma velocidade do rio e do tronco. Este terá de ir para o fim da lista até atingir as coordenadas (0,0).

Exemplos de utilização:

>>> movimentoNoTronco (Jogador (2,-2)) (Mapa 10 [(Rio 2, [Nenhum])])
Jogador (2,-2)

maybeMove :: Largura -> Coordenadas -> Jogada -> Bool

Função 4

A função maybeMove tendo a largura do mapa quando o personagem se encontra nas extremidades do mapa, não sofre qualquer efeito na sua posição. Esta função é importante para o movimento do Jogador.

Exemplos de utilização:

>>> maybeMove 10 (2,-2) (Move Baixo)
True

movimentoObstaculos :: [(Terreno, [Obstaculo])] -> [(Terreno, [Obstaculo])]

Função 5

A função movimentosObstaculos ao deslocar os obstáculos de uma linha, temos de ter em conta que ao desaparecerem por um dos lados do mapa devem reaparecer no lado oposto.

Exemplos de utilização:

>>> movimentoObstaculos 10 [Nenhum,Tronco,Nenhum]
[Nenhum,Nenhum,Tronco]

>>> movimentoObstaculos -10 [Nenhum,Tronco,Tronco]
[Tronco,Tronco,Nenhum]

Propriedades:

movimentoObstaculos [] = []

movimentoEsquerda :: Int -> [Obstaculo] -> [Obstaculo]

movimentoDireita :: Int -> [Obstaculo] -> [Obstaculo]