Tarefa4_2022li1g020

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2022/23.

Documentation

jogoTerminou :: Jogo -> Bool

Função Principal

A função jogoTerminou que serve para ver se o jogador perdeu, onde o resultado True significa que sim. Para isso usamos funções auxiliares para ver se o jogador se encontra fora do mapa, na água, ou por baixo de um carro, isto é, se encontre na mesma posição de um carro.

Exemplos de utilização:

>>> jogoTerminou (Jogo (Jogador (2,-3)) (Mapa 10 [(Relva,[Arvore,Nenhum,Nenhum]),(Estrada 2,[Carro,Nenhum])]))
False

Propriedades:

>>> jogoTerminou (Jogo (Jogador (x,y)) (Mapa larg [])) = False

morreste :: Jogador -> [(Terreno, [Obstaculo])] -> Bool

O jogo termina se o jogador morreu, se o jogador saiu do mapa ou se o jogador foi atropelado por um veículo. Essas condições são verificadas pelas funções morreste, foramapa e atropelado, respectivamente.

A função morreste recebe um jogador Jogador e a lista de linhas do mapa t, e retorna um valor booleano indicando se o jogador morreu ou não. Ela faz isso verificando se o jogador está na água (se estiver, ele morre) ou se ele está em uma linha de estrada e há um obstáculo (carro) na mesma posição do jogador.

morreste1 :: Jogador -> (Terreno, [Obstaculo]) -> Bool

xNenhum4 :: Obstaculo -> [Obstaculo] -> [Int]

A função foramapa recebe a largura do mapa l, o jogador Jogador e a lista de linhas do mapa t, e retorna um valor booleano indicando se o jogador saiu do mapa ou não.

A função atropelado recebe o jogador Jogador e o mapa Mapa, e retorna um valor booleano indicando se o jogador foi atropelado ou não. Ela faz isso verificando se o jogador está em uma linha de estrada e há um veículo se movendo na mesma direção e passando pelo jogador.

As funções xNenhum4 e xNenhum4aux são auxiliares na verificação de se o jogador morreu ou foi atropelado. A função xNenhum4 recebe um tipo de obstáculo Obstaculo e uma lista de obstáculos [Obstaculo], e retorna uma lista de inteiros indicando as posições na lista de obstáculos em que o obstáculo especificado aparece. A função xNenhum4aux é uma função recursiva auxiliar de xNenhum4, que percorre a lista de obstáculos ver

xNenhum4aux :: Obstaculo -> [Obstaculo] -> Int -> [Int]

foramapa :: Largura -> Jogador -> [(Terreno, [Obstaculo])] -> Bool

atropelado :: Jogador -> Mapa -> Bool

atropeladoaux :: Jogador -> (Terreno, [Obstaculo]) -> Bool

testa :: Int -> Int -> [Int] -> Bool

move4 :: Jogador -> Jogada -> Jogador

As funções move4 e movelinha4 são responsáveis por mover os obstáculos (veículos) do mapa. A função move4 recebe um jogador Jogador e uma jogada Jogada e retorna um novo jogador resultante da aplicação da jogada ao jogador. A função movelinha4 recebe um mapa Mapa e retorna um novo mapa com os obstáculos das linhas do mapa movidos de acordo com as respectivas velocidades.

movelinha4 :: Mapa -> Mapa

movelinhaaux4 :: [(Terreno, [Obstaculo])] -> [(Terreno, [Obstaculo])]

deslocaEsq :: Int -> [a] -> [a]

deslocaDireita :: Int -> [a] -> [a]