Tarefa5_2022li1g020

Description

Módulo para a realização da Tarefa 5 do projeto de LI1 em 2022/23.

Documentation

deslizaJogo :: Int -> Jogo -> Jogo

Função Principal

Expand

A função deslizaJogo faz com que a última linha do mapa desapareça e apareça uma nova linha na parte superior do mapa, de forma de continuar o jogo. Desta maneira, o mapa mantêm o tamanho original. A nova linha adicionada é criada de uma forma aleatória, através do número inteiro apresentado, que passará pela função estendeMapa anteriormente definida. Temos também de ter atenção à adição de uma unidade no eixo dos yy, para o jogador não ser afetado por esta alteraçãop e se manter no mesmo sítio.

remove :: Mapa -> Mapa